Fix default Max Maturity value type
Since this is an U32, I think it should be better to default it to an integer instead of a boolean.
I have set it to 13, which is the value for SIM_ACCESS_PG.

Perhaps should we change it to U8 if the settings system supports it to avoid needless typecasts?